Database/passing data update
On the second screen of the farmer sign up, which is as far as that sequence goes, I insert the data into mySQL database. I have passed all data through the farmer request transport sequence and am ready to finish backend integration once the transporter sign up sequence is finished.